Problem: weird memory leaks reported by valgrind
Solution: suppress them

It is NOT clear why this is happening, but it happens in
locale_facets.tcc, which build some cache for locale conversions on
demand. Given the fact objects are allocated inside libstc++, it sounds
that not destroying objects is harmless, as they will be free'ed whn
program ends.

Plus the number of leaks does not increate with number of test cases
(and save_alerts_state is called almost each time).

Signed-off-by: Michal Vyskocil <michalvyskocil@eaton.com>